Show coordinates on map

+++ Earthquake Monitor/Earthquake.swift
@@ -13,7 +13,7 @@ class Earthquake {
     var place: String?
     var mag: Float?
     var time: Double?
-    var longitude: Float?
-    var latitude: Float?
+    var longitude: Double?
+    var latitude: Double?
     var depth: Float?
 }
